requirements.txt
aiomysql==0.1.1
aiosqlite==0.17.0
anyio==3.6.1
autopep8==1.7.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.1
h11==0.13.0
idna==3.3
iso8601==1.0.2
passlib==1.7.4
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.2
PyJWT==2.4.0
PyMySQL==1.0.2
pypika-tortoise==0.1.6
python-dotenv==0.20.0
pytz==2022.2.1
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.2
